missed a long to int change

+++ src/test/java/org/folio/edge/rtac/MainVerticleTest.java
@@ -64,7 +64,7 @@ public class MainVerticleTest {
   private static final String badApiKey = apiKey + "0000";
   private static final String unknownTenantApiKey = ApiKeyUtils.generateApiKey(10, "bogus", "diku");
 
-  private static final long requestTimeoutMs = 3000L;
+  private static final int requestTimeoutMs = 3000;
 
   private static Vertx vertx;
   protected static RtacMockOkapi mockOkapi;
